Keep point in timeline in sync with the timemachine
When going to the next or previous snapshot in a timemachine buffer while the
timeline is displayed in another window, the point would not actually move to
the correct snapshot in the timeline. This is annoying and confusing as after
switching back to the timeline buffer, another snapshot would appear to be
selected.

Fix this by using `with-selected-window` instead of `with-current-buffer`.